requirements.txt
pandas
numpy
scipy
matplotlib
keras
jupyter
GitPython
